Update README
Extended on the vocabulary usage to provide a more streamlined user interface; basically this will change how the reader will experience the words he or she is reading in such a way as to elicit some manner of positive emotional response. Hopefully nothing bad will happen due to the capitalization of certain words being correlated with high suicide rates in certain countries.